+++ Source/Object/World/World.h
@@ -1,7 +1,13 @@
 #pragma once
 #include "JsonSavehelper.h"
+#include "Core/Engine.h"
+#include "Core/Container/Array.h"
 #include "Object/UObject.h"
-#include "Core/Container/Set.h"
+#include "Debug/DebugConsole.h"
+#include "Object/ObjectFactory.h"
+
+class AActor;
+
 
 class UWorld :public UObject
 {
@@ -16,12 +22,32 @@ public:
 	void SaveWorld();
 	void BeginPlay();
 	void Tick(float DeltaTime);
-	void AddActor(class AActor* Actor) { Actors.Add(Actor); }
+
+	template <typename T>
+		requires std::derived_from<T, AActor>
+	T* SpawnActor();
 
 public:
-	std::string SceneName = "";
+	std::string SceneName;
 	
 protected:
-	TSet<class AActor*> Actors;
+	TArray<AActor*> Actors;
 };
+
+template <typename T>
+	requires std::derived_from<T, AActor>
+T* UWorld::SpawnActor()
+{
+	T* Actor = FObjectFactory::ConstructObject<T>();
+	
+	if (UWorld* World = UEngine::Get().GetWorld())
+	{
+		Actor->SetWorld(World);
+		Actors.Add(Actor);
+		return Actor;
+	}
+
+	UE_LOG("Actor Construction Failed. World is nullptr");
+	return nullptr;
+}
